Resolve unresolved references (#6)
* Added support for resolving missing references to people on the "primary relationships" used for authorization.

This change resolves a situation that can occur when 1) resources are published, 2) a studentSchoolAssociation is added for an existing student who has not been updated, 3) data is published again. The student will not be available in the target because it was not available until the studentSchoolAssociation was added, and has not been subsequently updated in the source. This change will attempt to move the resource for the unresolved reference from the source API to the target API and retry.

* Improved logging around missing reference resolution.

* Upgraded log4net version, and added reference to AWS logger for log4net.

* Fixed problem related to posting the wrong item for the unresolved reference to the target API.

* Added the response content to the log for the failed POST request to the target API's resource for the unresolved reference.

* Added test coverage for the resolve references behavior to ensure that the item retrieved from the source API for the unresolved reference on the target is included in the POST request against the target.